Computes the actual and predicted effectiveness of a given algorithm.

This function computes the actual and predicted effectiveness of a given algorithm for different tolerance values.

algo_effectiveness_crm(mod, num = 1)

Arguments

mod
A fitted mirt model using the function irtmodel or R package mirt.

num
The algorithm number, for which the goodness of the IRT model is computed.

Value

A list with the following components:

effective

The x,y coodinates for the actual and predicted effectiveness curves for algorithm num.

predictedEff

The area under the predicted effectiveness curve.

actualEff

The area under the actual effectiveness curve.

Examples

# \donttest{
set.seed(1)
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)

#> Warning: NaNs produced

#> Warning: NaNs produced

out <- algo_effectiveness_crm(mod$model, num=1)
out

#> $effective
#>                 x act_rel pred_rel
#>   [1,] 0.00000000    0.00     0.00
#>   [2,] 0.01010101    0.01     0.00
#>   [3,] 0.02020202    0.01     0.00
#>   [4,] 0.03030303    0.01     0.01
#>   [5,] 0.04040404    0.02     0.01
#>   [6,] 0.05050505    0.02     0.01
#>   [7,] 0.06060606    0.03     0.01
#>   [8,] 0.07070707    0.04     0.02
#>   [9,] 0.08080808    0.04     0.02
#>  [10,] 0.09090909    0.05     0.02
#>  [11,] 0.10101010    0.06     0.02
#>  [12,] 0.11111111    0.08     0.02
#>  [13,] 0.12121212    0.08     0.03
#>  [14,] 0.13131313    0.10     0.03
#>  [15,] 0.14141414    0.13     0.04
#>  [16,] 0.15151515    0.13     0.06
#>  [17,] 0.16161616    0.13     0.06
#>  [18,] 0.17171717    0.14     0.07
#>  [19,] 0.18181818    0.16     0.07
#>  [20,] 0.19191919    0.17     0.07
#>  [21,] 0.20202020    0.17     0.10
#>  [22,] 0.21212121    0.19     0.10
#>  [23,] 0.22222222    0.21     0.12
#>  [24,] 0.23232323    0.24     0.13
#>  [25,] 0.24242424    0.26     0.16
#>  [26,] 0.25252525    0.27     0.17
#>  [27,] 0.26262626    0.27     0.17
#>  [28,] 0.27272727    0.28     0.22
#>  [29,] 0.28282828    0.29     0.23
#>  [30,] 0.29292929    0.32     0.26
#>  [31,] 0.30303030    0.32     0.28
#>  [32,] 0.31313131    0.33     0.30
#>  [33,] 0.32323232    0.34     0.31
#>  [34,] 0.33333333    0.34     0.33
#>  [35,] 0.34343434    0.37     0.33
#>  [36,] 0.35353535    0.39     0.34
#>  [37,] 0.36363636    0.41     0.36
#>  [38,] 0.37373737    0.41     0.37
#>  [39,] 0.38383838    0.42     0.37
#>  [40,] 0.39393939    0.42     0.39
#>  [41,] 0.40404040    0.43     0.40
#>  [42,] 0.41414141    0.44     0.41
#>  [43,] 0.42424242    0.44     0.43
#>  [44,] 0.43434343    0.45     0.44
#>  [45,] 0.44444444    0.45     0.46
#>  [46,] 0.45454545    0.46     0.48
#>  [47,] 0.46464646    0.46     0.50
#>  [48,] 0.47474747    0.46     0.50
#>  [49,] 0.48484848    0.47     0.50
#>  [50,] 0.49494949    0.48     0.51
#>  [51,] 0.50505051    0.48     0.52
#>  [52,] 0.51515152    0.50     0.54
#>  [53,] 0.52525253    0.51     0.56
#>  [54,] 0.53535354    0.55     0.56
#>  [55,] 0.54545455    0.55     0.57
#>  [56,] 0.55555556    0.57     0.61
#>  [57,] 0.56565657    0.57     0.64
#>  [58,] 0.57575758    0.59     0.65
#>  [59,] 0.58585859    0.59     0.66
#>  [60,] 0.59595960    0.61     0.66
#>  [61,] 0.60606061    0.62     0.69
#>  [62,] 0.61616162    0.63     0.69
#>  [63,] 0.62626263    0.68     0.70
#>  [64,] 0.63636364    0.69     0.72
#>  [65,] 0.64646465    0.69     0.72
#>  [66,] 0.65656566    0.69     0.74
#>  [67,] 0.66666667    0.71     0.76
#>  [68,] 0.67676768    0.74     0.77
#>  [69,] 0.68686869    0.75     0.79
#>  [70,] 0.69696970    0.76     0.82
#>  [71,] 0.70707071    0.76     0.83
#>  [72,] 0.71717172    0.77     0.84
#>  [73,] 0.72727273    0.77     0.85
#>  [74,] 0.73737374    0.77     0.85
#>  [75,] 0.74747475    0.79     0.85
#>  [76,] 0.75757576    0.80     0.87
#>  [77,] 0.76767677    0.82     0.87
#>  [78,] 0.77777778    0.83     0.87
#>  [79,] 0.78787879    0.83     0.89
#>  [80,] 0.79797980    0.84     0.92
#>  [81,] 0.80808081    0.87     0.93
#>  [82,] 0.81818182    0.87     0.94
#>  [83,] 0.82828283    0.88     0.95
#>  [84,] 0.83838384    0.89     0.95
#>  [85,] 0.84848485    0.89     0.96
#>  [86,] 0.85858586    0.89     0.96
#>  [87,] 0.86868687    0.90     0.96
#>  [88,] 0.87878788    0.90     0.97
#>  [89,] 0.88888889    0.92     0.97
#>  [90,] 0.89898990    0.92     0.97
#>  [91,] 0.90909091    0.93     0.98
#>  [92,] 0.91919192    0.94     0.98
#>  [93,] 0.92929293    0.95     0.98
#>  [94,] 0.93939394    0.96     0.99
#>  [95,] 0.94949495    0.97     0.99
#>  [96,] 0.95959596    0.98     1.00
#>  [97,] 0.96969697    0.98     1.00
#>  [98,] 0.97979798    0.98     1.00
#>  [99,] 0.98989899    0.99     1.00
#> [100,] 1.00000000    1.00     1.00
#> 
#> $predictedEff
#> [1] 0.5174747
#> 
#> $actualEff
#> [1] 0.5128283
#> 

# }