Fits a continuous IRT model.

This function fits a continuous Item Response Theory (IRT) model to the algorithm performance data. The function EstCRMitem in the R package EstCRM is updated to accommodate negative discrimination.

cirtmodel(df, max.item = NULL, min.item = NULL)

Arguments

df

The performance data in a matrix or dataframe.

max.item

A vector with the maximum performance value for each algorithm.

min.item

A vector with the minimum performance value for each algorithm.

Value

A list with the following components:

model

The IRT model.

anomalous

A binary value for each algorithm. It is set to 1 if an algorithm is anomalous. Otherwise it is set to 0.

stability

The stability of each algorithm.

difficulty_limit

The difficulty limit of each algorithm. A higher difficulty limit indicates that the algorithm can tackle harder problems.

References

Zopluoglu C (2022). EstCRM: Calibrating Parameters for the Samejima's Continuous IRT Model. R package version 1.6, https://CRAN.R-project.org/package=EstCRM.

Examples

set.seed(1)
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)
#> Warning: NaNs produced
#> Warning: NaNs produced