Computes the goodness of IRT model for all algorithms.

This function computes the goodness of the IRT model for all algorithms for different goodness tolerances.

model_goodness_crm(mod)

Arguments

mod

A fitted mirt model using the function irtmodel or R package mirt.

Value

A list with the following components:

goodnessAUC

The area under the model goodness curve for each algorithm.

curves

The x,y coordinates for the model goodness curves for each algorithm.

residuals

The residuals for each algorithm using the AIRT model.

Examples

# \donttest{
set.seed(1)
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)
#> Warning: NaNs produced
#> Warning: NaNs produced
out <- model_goodness_crm(mod$model)
out
#> $goodnessAUC
#>         [,1]
#> x1 0.9253535
#> x2 0.5859596
#> x3 0.7127273
#> 
#> $mse
#>             [,1]
#> [1,] 0.008553221
#> [2,] 0.238807237
#> [3,] 0.125930464
#> 
#> $curves
#>                 x   x1   x2   x3
#>   [1,] 0.00000000 0.00 0.00 0.00
#>   [2,] 0.01010101 0.11 0.04 0.01
#>   [3,] 0.02020202 0.18 0.05 0.03
#>   [4,] 0.03030303 0.21 0.05 0.05
#>   [5,] 0.04040404 0.32 0.05 0.07
#>   [6,] 0.05050505 0.38 0.06 0.13
#>   [7,] 0.06060606 0.45 0.06 0.16
#>   [8,] 0.07070707 0.55 0.06 0.18
#>   [9,] 0.08080808 0.59 0.09 0.18
#>  [10,] 0.09090909 0.66 0.09 0.19
#>  [11,] 0.10101010 0.75 0.11 0.20
#>  [12,] 0.11111111 0.77 0.13 0.22
#>  [13,] 0.12121212 0.83 0.19 0.22
#>  [14,] 0.13131313 0.86 0.21 0.25
#>  [15,] 0.14141414 0.89 0.24 0.28
#>  [16,] 0.15151515 0.92 0.24 0.33
#>  [17,] 0.16161616 0.94 0.26 0.34
#>  [18,] 0.17171717 0.94 0.27 0.36
#>  [19,] 0.18181818 0.94 0.30 0.38
#>  [20,] 0.19191919 0.96 0.31 0.38
#>  [21,] 0.20202020 0.98 0.31 0.41
#>  [22,] 0.21212121 0.98 0.32 0.43
#>  [23,] 0.22222222 0.98 0.35 0.46
#>  [24,] 0.23232323 0.99 0.35 0.47
#>  [25,] 0.24242424 0.99 0.35 0.53
#>  [26,] 0.25252525 0.99 0.36 0.55
#>  [27,] 0.26262626 0.99 0.37 0.58
#>  [28,] 0.27272727 0.99 0.38 0.60
#>  [29,] 0.28282828 0.99 0.38 0.62
#>  [30,] 0.29292929 0.99 0.39 0.62
#>  [31,] 0.30303030 0.99 0.39 0.62
#>  [32,] 0.31313131 1.00 0.39 0.64
#>  [33,] 0.32323232 1.00 0.40 0.67
#>  [34,] 0.33333333 1.00 0.41 0.69
#>  [35,] 0.34343434 1.00 0.42 0.69
#>  [36,] 0.35353535 1.00 0.42 0.70
#>  [37,] 0.36363636 1.00 0.44 0.70
#>  [38,] 0.37373737 1.00 0.44 0.71
#>  [39,] 0.38383838 1.00 0.47 0.72
#>  [40,] 0.39393939 1.00 0.48 0.73
#>  [41,] 0.40404040 1.00 0.51 0.75
#>  [42,] 0.41414141 1.00 0.53 0.76
#>  [43,] 0.42424242 1.00 0.53 0.76
#>  [44,] 0.43434343 1.00 0.54 0.79
#>  [45,] 0.44444444 1.00 0.55 0.80
#>  [46,] 0.45454545 1.00 0.56 0.80
#>  [47,] 0.46464646 1.00 0.58 0.81
#>  [48,] 0.47474747 1.00 0.59 0.81
#>  [49,] 0.48484848 1.00 0.59 0.81
#>  [50,] 0.49494949 1.00 0.61 0.82
#>  [51,] 0.50505051 1.00 0.62 0.82
#>  [52,] 0.51515152 1.00 0.63 0.83
#>  [53,] 0.52525253 1.00 0.66 0.83
#>  [54,] 0.53535354 1.00 0.67 0.83
#>  [55,] 0.54545455 1.00 0.67 0.83
#>  [56,] 0.55555556 1.00 0.69 0.84
#>  [57,] 0.56565657 1.00 0.70 0.85
#>  [58,] 0.57575758 1.00 0.72 0.87
#>  [59,] 0.58585859 1.00 0.73 0.87
#>  [60,] 0.59595960 1.00 0.73 0.87
#>  [61,] 0.60606061 1.00 0.73 0.87
#>  [62,] 0.61616162 1.00 0.73 0.88
#>  [63,] 0.62626263 1.00 0.74 0.90
#>  [64,] 0.63636364 1.00 0.74 0.90
#>  [65,] 0.64646465 1.00 0.74 0.90
#>  [66,] 0.65656566 1.00 0.75 0.91
#>  [67,] 0.66666667 1.00 0.76 0.91
#>  [68,] 0.67676768 1.00 0.78 0.92
#>  [69,] 0.68686869 1.00 0.81 0.93
#>  [70,] 0.69696970 1.00 0.81 0.93
#>  [71,] 0.70707071 1.00 0.82 0.93
#>  [72,] 0.71717172 1.00 0.82 0.93
#>  [73,] 0.72727273 1.00 0.85 0.94
#>  [74,] 0.73737374 1.00 0.86 0.94
#>  [75,] 0.74747475 1.00 0.86 0.97
#>  [76,] 0.75757576 1.00 0.86 0.98
#>  [77,] 0.76767677 1.00 0.86 0.98
#>  [78,] 0.77777778 1.00 0.87 0.98
#>  [79,] 0.78787879 1.00 0.89 0.98
#>  [80,] 0.79797980 1.00 0.90 0.98
#>  [81,] 0.80808081 1.00 0.90 0.98
#>  [82,] 0.81818182 1.00 0.91 0.98
#>  [83,] 0.82828283 1.00 0.91 0.99
#>  [84,] 0.83838384 1.00 0.92 1.00
#>  [85,] 0.84848485 1.00 0.92 1.00
#>  [86,] 0.85858586 1.00 0.92 1.00
#>  [87,] 0.86868687 1.00 0.93 1.00
#>  [88,] 0.87878788 1.00 0.95 1.00
#>  [89,] 0.88888889 1.00 0.96 1.00
#>  [90,] 0.89898990 1.00 0.97 1.00
#>  [91,] 0.90909091 1.00 1.00 1.00
#>  [92,] 0.91919192 1.00 1.00 1.00
#>  [93,] 0.92929293 1.00 1.00 1.00
#>  [94,] 0.93939394 1.00 1.00 1.00
#>  [95,] 0.94949495 1.00 1.00 1.00
#>  [96,] 0.95959596 1.00 1.00 1.00
#>  [97,] 0.96969697 1.00 1.00 1.00
#>  [98,] 0.97979798 1.00 1.00 1.00
#>  [99,] 0.98989899 1.00 1.00 1.00
#> [100,] 1.00000000 1.00 1.00 1.00
#> 
#> $residuals
#>                  x1           x2         x3
#>   [1,] 6.195397e-02 0.1374180296 0.24035534
#>   [2,] 2.722307e-03 0.3717734244 0.17805264
#>   [3,] 1.819791e-02 0.7201817119 0.32918738
#>   [4,] 5.108601e-02 0.0072705410 0.61915363
#>   [5,] 5.666190e-02 0.4599812947 0.16820113
#>   [6,] 8.924807e-02 0.7867464669 0.26564853
#>   [7,] 7.858722e-02 0.8705822541 0.34106724
#>   [8,] 1.169652e-01 0.5100840199 0.04391925
#>   [9,] 2.273938e-02 0.0725086154 0.01508255
#>  [10,] 1.189823e-01 0.5765180381 0.71325409
#>  [11,] 3.087715e-01 0.0031865317 0.82278185
#>  [12,] 4.553661e-02 0.6579535443 0.09181335
#>  [13,] 2.190798e-03 0.6426308303 0.31636107
#>  [14,] 1.859441e-01 0.5614591889 0.73796036
#>  [15,] 3.321780e-02 0.8516003803 0.23090508
#>  [16,] 9.405361e-02 0.8288379390 0.25296849
#>  [17,] 9.502097e-03 0.2840581839 0.14094182
#>  [18,] 1.716484e-02 0.8967703664 0.04473624
#>  [19,] 1.938225e-01 0.5471606998 0.73785879
#>  [20,] 7.940207e-02 0.3594237644 0.14152383
#>  [21,] 5.166646e-02 0.0081159821 0.66766810
#>  [22,] 4.233431e-02 0.3357862619 0.15305801
#>  [23,] 6.205100e-02 0.5108601933 0.08059652
#>  [24,] 8.765912e-02 0.1151927577 0.50271284
#>  [25,] 2.225464e-01 0.2120314229 0.81007671
#>  [26,] 5.796657e-02 0.4687614954 0.43260126
#>  [27,] 3.661901e-02 0.5111485282 0.25743421
#>  [28,] 8.892626e-02 0.1232863408 0.02731561
#>  [29,] 1.026797e-01 0.7712567967 0.25505271
#>  [30,] 1.597273e-01 0.3770048811 0.73040411
#>  [31,] 5.078325e-03 0.3903123061 0.23568812
#>  [32,] 1.431768e-01 0.8611937682 0.04461275
#>  [33,] 6.081244e-02 0.8695034024 0.30330047
#>  [34,] 1.356261e-01 0.1566653187 0.60574971
#>  [35,] 5.246896e-02 0.0713163722 0.30457174
#>  [36,] 5.592806e-03 0.4010906230 0.27008537
#>  [37,] 3.575699e-02 0.4389922601 0.04025803
#>  [38,] 1.129029e-01 0.4545601586 0.54670643
#>  [39,] 5.125078e-03 0.0146765613 0.13626793
#>  [40,] 4.350299e-02 0.4287585628 0.36704268
#>  [41,] 8.903846e-02 0.3170003893 0.26805007
#>  [42,] 6.639881e-03 0.3967696122 0.23780068
#>  [43,] 2.182075e-03 0.7610732832 0.19981682
#>  [44,] 9.702728e-02 0.6796589129 0.05604737
#>  [45,] 1.489959e-02 0.2590710477 0.24094794
#>  [46,] 1.194763e-01 0.5465332927 0.19826418
#>  [47,] 4.907698e-02 0.1749821415 0.42829662
#>  [48,] 3.493994e-02 0.1734004471 0.05550245
#>  [49,] 9.513485e-02 0.8932038580 0.12660547
#>  [50,] 8.877630e-02 0.1353982496 0.23035851
#>  [51,] 9.074013e-02 0.3780488407 0.55452605
#>  [52,] 4.237290e-03 0.4427950633 0.04507889
#>  [53,] 2.405924e-02 0.6157139127 0.37961538
#>  [54,] 1.107337e-01 0.2099500864 0.56794110
#>  [55,] 9.363260e-02 0.4876313511 0.48137462
#>  [56,] 3.470656e-02 0.1768102434 0.10668727
#>  [57,] 3.472804e-02 0.1115561031 0.21358248
#>  [58,] 3.177906e-02 0.8897096802 0.38724232
#>  [59,] 7.147314e-02 0.7175402206 0.12974235
#>  [60,] 1.487777e-01 0.7779184245 0.74575581
#>  [61,] 8.231067e-02 0.7151641212 0.31041482
#>  [62,] 9.088787e-02 0.1122997598 0.27053222
#>  [63,] 1.141222e-02 0.4812976317 0.19722538
#>  [64,] 1.576012e-01 0.1870099820 0.67344072
#>  [65,] 4.044265e-02 0.1188264523 0.23875259
#>  [66,] 2.462302e-02 0.1365080009 0.14798938
#>  [67,] 1.370820e-01 0.5289632721 0.03330508
#>  [68,] 3.017865e-02 0.6643221173 0.16835728
#>  [69,] 1.252158e-01 0.6712036171 0.61336113
#>  [70,] 1.369544e-01 0.6622046709 0.42305318
#>  [71,] 3.030335e-02 0.2171049488 0.03974032
#>  [72,] 1.232899e-01 0.1590749355 0.42161627
#>  [73,] 8.269082e-02 0.0430341233 0.31253816
#>  [74,] 1.639224e-06 0.1604670597 0.14041084
#>  [75,] 1.784593e-02 0.5601253811 0.21028964
#>  [76,] 9.217476e-02 0.1045091761 0.56911569
#>  [77,] 1.021657e-01 0.3555872892 0.40229269
#>  [78,] 1.070368e-02 0.0927180348 0.12733434
#>  [79,] 6.017483e-02 0.3943933618 0.05607775
#>  [80,] 3.050217e-02 0.0968729906 0.20871994
#>  [81,] 1.240962e-01 0.1295166010 0.00489488
#>  [82,] 6.697123e-02 0.8072243404 0.17539192
#>  [83,] 1.834339e-01 0.1081280087 0.64912030
#>  [84,] 5.511523e-02 0.2650248826 0.32812508
#>  [85,] 1.476570e-01 0.1197315654 0.22923238
#>  [86,] 1.845915e-02 0.1171518376 0.13040160
#>  [87,] 5.772288e-02 0.2406154071 0.04832942
#>  [88,] 6.761585e-02 0.6954011291 0.14951763
#>  [89,] 6.797183e-02 0.5043666416 0.10864678
#>  [90,] 6.942158e-02 0.4901699776 0.21394163
#>  [91,] 6.385378e-02 0.3274693843 0.20510489
#>  [92,] 4.812295e-02 0.3877384793 0.13022582
#>  [93,] 9.505443e-02 0.8989421794 0.39181297
#>  [94,] 6.734479e-02 0.0726525135 0.45667524
#>  [95,] 7.710743e-02 0.7163319887 0.01536322
#>  [96,] 3.610332e-02 0.4093273433 0.02468031
#>  [97,] 9.828169e-02 0.5596167983 0.06302583
#>  [98,] 6.636813e-03 0.0008738496 0.06367554
#>  [99,] 1.978298e-01 0.6790536300 0.23365275
#> [100,] 1.129333e-01 0.2168506723 0.39387180
#> 
# }