Computes the goodness of the IRT model fit for a given algorithm.

This function computes the goodness of the IRT model fit for a given algorithm using the empirical cumulative distribution function of errors.

model_goodness_for_algo_poly(mod, num = 1)

Arguments

mod
A fitted mirt model using the function irtmodel or R package mirt.

num
The algorithm number

Value

A list with the following components:

xy

The x values denote the error tolerances. The y values denotes its empirical cumulative distribution function.

auc

The area under the CDF.

mse

The mean squared error.

Examples

set.seed(1)
x1 <- sample(1:5, 100, replace = TRUE)
x2 <- sample(1:5, 100, replace = TRUE)
x3 <- sample(1:5, 100, replace = TRUE)
X <- cbind.data.frame(x1, x2, x3)
mod <- pirtmodel(X)

#> Iteration: 1, Log-Lik: -515.114, Max-Change: 0.41351Iteration: 2, Log-Lik: -489.992, Max-Change: 0.19653Iteration: 3, Log-Lik: -483.285, Max-Change: 0.10168Iteration: 4, Log-Lik: -481.150, Max-Change: 0.06167Iteration: 5, Log-Lik: -480.335, Max-Change: 0.04035Iteration: 6, Log-Lik: -479.979, Max-Change: 0.02866Iteration: 7, Log-Lik: -479.740, Max-Change: 0.01807Iteration: 8, Log-Lik: -479.673, Max-Change: 0.01388Iteration: 9, Log-Lik: -479.634, Max-Change: 0.01086Iteration: 10, Log-Lik: -479.601, Max-Change: 0.00763Iteration: 11, Log-Lik: -479.588, Max-Change: 0.00611Iteration: 12, Log-Lik: -479.579, Max-Change: 0.00555Iteration: 13, Log-Lik: -479.562, Max-Change: 0.00850Iteration: 14, Log-Lik: -479.557, Max-Change: 0.00404Iteration: 15, Log-Lik: -479.553, Max-Change: 0.00415Iteration: 16, Log-Lik: -479.533, Max-Change: 0.00412Iteration: 17, Log-Lik: -479.529, Max-Change: 0.00421Iteration: 18, Log-Lik: -479.525, Max-Change: 0.00441Iteration: 19, Log-Lik: -479.516, Max-Change: 0.00449Iteration: 20, Log-Lik: -479.512, Max-Change: 0.00436Iteration: 21, Log-Lik: -479.507, Max-Change: 0.00386Iteration: 22, Log-Lik: -479.497, Max-Change: 0.00384Iteration: 23, Log-Lik: -479.493, Max-Change: 0.00372Iteration: 24, Log-Lik: -479.489, Max-Change: 0.00360Iteration: 25, Log-Lik: -479.480, Max-Change: 0.00348Iteration: 26, Log-Lik: -479.476, Max-Change: 0.00330Iteration: 27, Log-Lik: -479.473, Max-Change: 0.00329Iteration: 28, Log-Lik: -479.461, Max-Change: 0.00330Iteration: 29, Log-Lik: -479.458, Max-Change: 0.00321Iteration: 30, Log-Lik: -479.456, Max-Change: 0.00318Iteration: 31, Log-Lik: -479.441, Max-Change: 0.00294Iteration: 32, Log-Lik: -479.439, Max-Change: 0.00259Iteration: 33, Log-Lik: -479.438, Max-Change: 0.00341Iteration: 34, Log-Lik: -479.435, Max-Change: 0.00282Iteration: 35, Log-Lik: -479.433, Max-Change: 0.00271Iteration: 36, Log-Lik: -479.432, Max-Change: 0.00255Iteration: 37, Log-Lik: -479.430, Max-Change: 0.00288Iteration: 38, Log-Lik: -479.429, Max-Change: 0.00270Iteration: 39, Log-Lik: -479.428, Max-Change: 0.00246Iteration: 40, Log-Lik: -479.426, Max-Change: 0.00236Iteration: 41, Log-Lik: -479.425, Max-Change: 0.00247Iteration: 42, Log-Lik: -479.424, Max-Change: 0.00231Iteration: 43, Log-Lik: -479.423, Max-Change: 0.00235Iteration: 44, Log-Lik: -479.422, Max-Change: 0.00217Iteration: 45, Log-Lik: -479.421, Max-Change: 0.00209Iteration: 46, Log-Lik: -479.417, Max-Change: 0.00258Iteration: 47, Log-Lik: -479.417, Max-Change: 0.00203Iteration: 48, Log-Lik: -479.416, Max-Change: 0.00197Iteration: 49, Log-Lik: -479.413, Max-Change: 0.00163Iteration: 50, Log-Lik: -479.413, Max-Change: 0.00171Iteration: 51, Log-Lik: -479.412, Max-Change: 0.00170Iteration: 52, Log-Lik: -479.410, Max-Change: 0.00191Iteration: 53, Log-Lik: -479.409, Max-Change: 0.00182Iteration: 54, Log-Lik: -479.409, Max-Change: 0.00162Iteration: 55, Log-Lik: -479.407, Max-Change: 0.00156Iteration: 56, Log-Lik: -479.407, Max-Change: 0.00148Iteration: 57, Log-Lik: -479.407, Max-Change: 0.00257Iteration: 58, Log-Lik: -479.406, Max-Change: 0.00150Iteration: 59, Log-Lik: -479.406, Max-Change: 0.00221Iteration: 60, Log-Lik: -479.406, Max-Change: 0.00139Iteration: 61, Log-Lik: -479.406, Max-Change: 0.00183Iteration: 62, Log-Lik: -479.406, Max-Change: 0.00135Iteration: 63, Log-Lik: -479.405, Max-Change: 0.00259Iteration: 64, Log-Lik: -479.405, Max-Change: 0.00138Iteration: 65, Log-Lik: -479.405, Max-Change: 0.00218Iteration: 66, Log-Lik: -479.405, Max-Change: 0.00130Iteration: 67, Log-Lik: -479.404, Max-Change: 0.00173Iteration: 68, Log-Lik: -479.404, Max-Change: 0.00127Iteration: 69, Log-Lik: -479.404, Max-Change: 0.00246Iteration: 70, Log-Lik: -479.404, Max-Change: 0.00129Iteration: 71, Log-Lik: -479.404, Max-Change: 0.00208Iteration: 72, Log-Lik: -479.403, Max-Change: 0.00122Iteration: 73, Log-Lik: -479.403, Max-Change: 0.00166Iteration: 74, Log-Lik: -479.403, Max-Change: 0.00119Iteration: 75, Log-Lik: -479.403, Max-Change: 0.00234Iteration: 76, Log-Lik: -479.403, Max-Change: 0.00122Iteration: 77, Log-Lik: -479.403, Max-Change: 0.00199Iteration: 78, Log-Lik: -479.402, Max-Change: 0.00115Iteration: 79, Log-Lik: -479.402, Max-Change: 0.00158Iteration: 80, Log-Lik: -479.402, Max-Change: 0.00112Iteration: 81, Log-Lik: -479.402, Max-Change: 0.00228Iteration: 82, Log-Lik: -479.402, Max-Change: 0.00115Iteration: 83, Log-Lik: -479.402, Max-Change: 0.00188Iteration: 84, Log-Lik: -479.402, Max-Change: 0.00108Iteration: 85, Log-Lik: -479.401, Max-Change: 0.00103Iteration: 86, Log-Lik: -479.401, Max-Change: 0.00246Iteration: 87, Log-Lik: -479.401, Max-Change: 0.00109Iteration: 88, Log-Lik: -479.401, Max-Change: 0.00105Iteration: 89, Log-Lik: -479.401, Max-Change: 0.00220Iteration: 90, Log-Lik: -479.401, Max-Change: 0.00105Iteration: 91, Log-Lik: -479.401, Max-Change: 0.00101Iteration: 92, Log-Lik: -479.401, Max-Change: 0.00225Iteration: 93, Log-Lik: -479.401, Max-Change: 0.00104Iteration: 94, Log-Lik: -479.400, Max-Change: 0.00100Iteration: 95, Log-Lik: -479.400, Max-Change: 0.00218Iteration: 96, Log-Lik: -479.400, Max-Change: 0.00102Iteration: 97, Log-Lik: -479.400, Max-Change: 0.00098Iteration: 98, Log-Lik: -479.400, Max-Change: 0.00216Iteration: 99, Log-Lik: -479.400, Max-Change: 0.00100Iteration: 100, Log-Lik: -479.400, Max-Change: 0.00096Iteration: 101, Log-Lik: -479.400, Max-Change: 0.00212Iteration: 102, Log-Lik: -479.400, Max-Change: 0.00098Iteration: 103, Log-Lik: -479.400, Max-Change: 0.00095Iteration: 104, Log-Lik: -479.400, Max-Change: 0.00209Iteration: 105, Log-Lik: -479.399, Max-Change: 0.00096Iteration: 106, Log-Lik: -479.399, Max-Change: 0.00093Iteration: 107, Log-Lik: -479.399, Max-Change: 0.00206Iteration: 108, Log-Lik: -479.399, Max-Change: 0.00094Iteration: 109, Log-Lik: -479.399, Max-Change: 0.00090Iteration: 110, Log-Lik: -479.399, Max-Change: 0.00207Iteration: 111, Log-Lik: -479.399, Max-Change: 0.00093Iteration: 112, Log-Lik: -479.399, Max-Change: 0.00089Iteration: 113, Log-Lik: -479.399, Max-Change: 0.00201Iteration: 114, Log-Lik: -479.399, Max-Change: 0.00091Iteration: 115, Log-Lik: -479.399, Max-Change: 0.00087Iteration: 116, Log-Lik: -479.399, Max-Change: 0.00199Iteration: 117, Log-Lik: -479.398, Max-Change: 0.00090Iteration: 118, Log-Lik: -479.398, Max-Change: 0.00086Iteration: 119, Log-Lik: -479.398, Max-Change: 0.00196Iteration: 120, Log-Lik: -479.398, Max-Change: 0.00088Iteration: 121, Log-Lik: -479.398, Max-Change: 0.00085Iteration: 122, Log-Lik: -479.398, Max-Change: 0.00193Iteration: 123, Log-Lik: -479.398, Max-Change: 0.00087Iteration: 124, Log-Lik: -479.398, Max-Change: 0.00083Iteration: 125, Log-Lik: -479.398, Max-Change: 0.00191Iteration: 126, Log-Lik: -479.398, Max-Change: 0.00085Iteration: 127, Log-Lik: -479.398, Max-Change: 0.00082Iteration: 128, Log-Lik: -479.398, Max-Change: 0.00188Iteration: 129, Log-Lik: -479.398, Max-Change: 0.00084Iteration: 130, Log-Lik: -479.397, Max-Change: 0.00081Iteration: 131, Log-Lik: -479.397, Max-Change: 0.00073Iteration: 132, Log-Lik: -479.397, Max-Change: 0.00249Iteration: 133, Log-Lik: -479.397, Max-Change: 0.00091Iteration: 134, Log-Lik: -479.397, Max-Change: 0.00078Iteration: 135, Log-Lik: -479.397, Max-Change: 0.00197Iteration: 136, Log-Lik: -479.397, Max-Change: 0.00084Iteration: 137, Log-Lik: -479.397, Max-Change: 0.00074Iteration: 138, Log-Lik: -479.397, Max-Change: 0.00215Iteration: 139, Log-Lik: -479.397, Max-Change: 0.00085Iteration: 140, Log-Lik: -479.397, Max-Change: 0.00074Iteration: 141, Log-Lik: -479.397, Max-Change: 0.00202Iteration: 142, Log-Lik: -479.396, Max-Change: 0.00082Iteration: 143, Log-Lik: -479.396, Max-Change: 0.00072Iteration: 144, Log-Lik: -479.396, Max-Change: 0.00203Iteration: 145, Log-Lik: -479.396, Max-Change: 0.00081Iteration: 146, Log-Lik: -479.396, Max-Change: 0.00071Iteration: 147, Log-Lik: -479.396, Max-Change: 0.00198Iteration: 148, Log-Lik: -479.396, Max-Change: 0.00080Iteration: 149, Log-Lik: -479.396, Max-Change: 0.00070Iteration: 150, Log-Lik: -479.396, Max-Change: 0.00196Iteration: 151, Log-Lik: -479.396, Max-Change: 0.00078Iteration: 152, Log-Lik: -479.396, Max-Change: 0.00069Iteration: 153, Log-Lik: -479.396, Max-Change: 0.00193Iteration: 154, Log-Lik: -479.396, Max-Change: 0.00077Iteration: 155, Log-Lik: -479.395, Max-Change: 0.00068Iteration: 156, Log-Lik: -479.395, Max-Change: 0.00068Iteration: 157, Log-Lik: -479.395, Max-Change: 0.00368Iteration: 158, Log-Lik: -479.395, Max-Change: 0.00091Iteration: 159, Log-Lik: -479.395, Max-Change: 0.00074Iteration: 160, Log-Lik: -479.395, Max-Change: 0.00150Iteration: 161, Log-Lik: -479.395, Max-Change: 0.00124Iteration: 162, Log-Lik: -479.395, Max-Change: 0.00354Iteration: 163, Log-Lik: -479.395, Max-Change: 0.00100Iteration: 164, Log-Lik: -479.395, Max-Change: 0.00079Iteration: 165, Log-Lik: -479.395, Max-Change: 0.00066Iteration: 166, Log-Lik: -479.394, Max-Change: 0.00351Iteration: 167, Log-Lik: -479.394, Max-Change: 0.00072Iteration: 168, Log-Lik: -479.394, Max-Change: 0.00064Iteration: 169, Log-Lik: -479.394, Max-Change: 0.00323Iteration: 170, Log-Lik: -479.394, Max-Change: 0.00069Iteration: 171, Log-Lik: -479.394, Max-Change: 0.00062Iteration: 172, Log-Lik: -479.394, Max-Change: 0.00316Iteration: 173, Log-Lik: -479.394, Max-Change: 0.00067Iteration: 174, Log-Lik: -479.394, Max-Change: 0.00061Iteration: 175, Log-Lik: -479.394, Max-Change: 0.00311Iteration: 176, Log-Lik: -479.393, Max-Change: 0.00066Iteration: 177, Log-Lik: -479.393, Max-Change: 0.00060Iteration: 178, Log-Lik: -479.393, Max-Change: 0.00308Iteration: 179, Log-Lik: -479.393, Max-Change: 0.00065Iteration: 180, Log-Lik: -479.393, Max-Change: 0.00059Iteration: 181, Log-Lik: -479.393, Max-Change: 0.00304Iteration: 182, Log-Lik: -479.393, Max-Change: 0.00064Iteration: 183, Log-Lik: -479.393, Max-Change: 0.00058Iteration: 184, Log-Lik: -479.393, Max-Change: 0.00300Iteration: 185, Log-Lik: -479.393, Max-Change: 0.00063Iteration: 186, Log-Lik: -479.393, Max-Change: 0.00057Iteration: 187, Log-Lik: -479.393, Max-Change: 0.00297Iteration: 188, Log-Lik: -479.392, Max-Change: 0.00062Iteration: 189, Log-Lik: -479.392, Max-Change: 0.00056Iteration: 190, Log-Lik: -479.392, Max-Change: 0.00294Iteration: 191, Log-Lik: -479.392, Max-Change: 0.00061Iteration: 192, Log-Lik: -479.392, Max-Change: 0.00055Iteration: 193, Log-Lik: -479.392, Max-Change: 0.00291Iteration: 194, Log-Lik: -479.392, Max-Change: 0.00060Iteration: 195, Log-Lik: -479.392, Max-Change: 0.00055Iteration: 196, Log-Lik: -479.392, Max-Change: 0.00288Iteration: 197, Log-Lik: -479.392, Max-Change: 0.00060Iteration: 198, Log-Lik: -479.392, Max-Change: 0.00054Iteration: 199, Log-Lik: -479.392, Max-Change: 0.00285Iteration: 200, Log-Lik: -479.392, Max-Change: 0.00059Iteration: 201, Log-Lik: -479.392, Max-Change: 0.00053Iteration: 202, Log-Lik: -479.391, Max-Change: 0.00282Iteration: 203, Log-Lik: -479.391, Max-Change: 0.00058Iteration: 204, Log-Lik: -479.391, Max-Change: 0.00053Iteration: 205, Log-Lik: -479.391, Max-Change: 0.00280Iteration: 206, Log-Lik: -479.391, Max-Change: 0.00057Iteration: 207, Log-Lik: -479.391, Max-Change: 0.00052Iteration: 208, Log-Lik: -479.391, Max-Change: 0.00278Iteration: 209, Log-Lik: -479.391, Max-Change: 0.00057Iteration: 210, Log-Lik: -479.391, Max-Change: 0.00052Iteration: 211, Log-Lik: -479.391, Max-Change: 0.00275Iteration: 212, Log-Lik: -479.391, Max-Change: 0.00056Iteration: 213, Log-Lik: -479.391, Max-Change: 0.00051Iteration: 214, Log-Lik: -479.391, Max-Change: 0.00273Iteration: 215, Log-Lik: -479.391, Max-Change: 0.00055Iteration: 216, Log-Lik: -479.391, Max-Change: 0.00051Iteration: 217, Log-Lik: -479.391, Max-Change: 0.00190Iteration: 218, Log-Lik: -479.390, Max-Change: 0.00051Iteration: 219, Log-Lik: -479.390, Max-Change: 0.00047Iteration: 220, Log-Lik: -479.390, Max-Change: 0.00190Iteration: 221, Log-Lik: -479.390, Max-Change: 0.00051Iteration: 222, Log-Lik: -479.390, Max-Change: 0.00046Iteration: 223, Log-Lik: -479.390, Max-Change: 0.00210Iteration: 224, Log-Lik: -479.390, Max-Change: 0.00049Iteration: 225, Log-Lik: -479.390, Max-Change: 0.00045Iteration: 226, Log-Lik: -479.390, Max-Change: 0.00195Iteration: 227, Log-Lik: -479.390, Max-Change: 0.00049Iteration: 228, Log-Lik: -479.390, Max-Change: 0.00045Iteration: 229, Log-Lik: -479.390, Max-Change: 0.00214Iteration: 230, Log-Lik: -479.390, Max-Change: 0.00056Iteration: 231, Log-Lik: -479.390, Max-Change: 0.00046Iteration: 232, Log-Lik: -479.390, Max-Change: 0.00016Iteration: 233, Log-Lik: -479.390, Max-Change: 0.00080Iteration: 234, Log-Lik: -479.390, Max-Change: 0.00074Iteration: 235, Log-Lik: -479.390, Max-Change: 0.00138Iteration: 236, Log-Lik: -479.390, Max-Change: 0.00047Iteration: 237, Log-Lik: -479.390, Max-Change: 0.00046Iteration: 238, Log-Lik: -479.390, Max-Change: 0.00293Iteration: 239, Log-Lik: -479.390, Max-Change: 0.00066Iteration: 240, Log-Lik: -479.389, Max-Change: 0.00054Iteration: 241, Log-Lik: -479.389, Max-Change: 0.00017Iteration: 242, Log-Lik: -479.389, Max-Change: 0.00016Iteration: 243, Log-Lik: -479.389, Max-Change: 0.00283Iteration: 244, Log-Lik: -479.389, Max-Change: 0.00052Iteration: 245, Log-Lik: -479.389, Max-Change: 0.00047Iteration: 246, Log-Lik: -479.389, Max-Change: 0.00050Iteration: 247, Log-Lik: -479.389, Max-Change: 0.00233Iteration: 248, Log-Lik: -479.389, Max-Change: 0.00065Iteration: 249, Log-Lik: -479.389, Max-Change: 0.00053Iteration: 250, Log-Lik: -479.389, Max-Change: 0.00016Iteration: 251, Log-Lik: -479.389, Max-Change: 0.00081Iteration: 252, Log-Lik: -479.389, Max-Change: 0.00073Iteration: 253, Log-Lik: -479.389, Max-Change: 0.00128Iteration: 254, Log-Lik: -479.389, Max-Change: 0.00046Iteration: 255, Log-Lik: -479.389, Max-Change: 0.00046Iteration: 256, Log-Lik: -479.389, Max-Change: 0.00021Iteration: 257, Log-Lik: -479.389, Max-Change: 0.00015Iteration: 258, Log-Lik: -479.389, Max-Change: 0.00075Iteration: 259, Log-Lik: -479.389, Max-Change: 0.00066Iteration: 260, Log-Lik: -479.389, Max-Change: 0.00018Iteration: 261, Log-Lik: -479.389, Max-Change: 0.00012Iteration: 262, Log-Lik: -479.389, Max-Change: 0.00251Iteration: 263, Log-Lik: -479.389, Max-Change: 0.00051Iteration: 264, Log-Lik: -479.389, Max-Change: 0.00046Iteration: 265, Log-Lik: -479.389, Max-Change: 0.00014Iteration: 266, Log-Lik: -479.389, Max-Change: 0.00311Iteration: 267, Log-Lik: -479.389, Max-Change: 0.00089Iteration: 268, Log-Lik: -479.389, Max-Change: 0.00050Iteration: 269, Log-Lik: -479.389, Max-Change: 0.00051Iteration: 270, Log-Lik: -479.389, Max-Change: 0.00047Iteration: 271, Log-Lik: -479.388, Max-Change: 0.00124Iteration: 272, Log-Lik: -479.388, Max-Change: 0.00045Iteration: 273, Log-Lik: -479.388, Max-Change: 0.00022Iteration: 274, Log-Lik: -479.388, Max-Change: 0.00020Iteration: 275, Log-Lik: -479.388, Max-Change: 0.00090Iteration: 276, Log-Lik: -479.388, Max-Change: 0.00145Iteration: 277, Log-Lik: -479.388, Max-Change: 0.00047Iteration: 278, Log-Lik: -479.388, Max-Change: 0.00043Iteration: 279, Log-Lik: -479.388, Max-Change: 0.00022Iteration: 280, Log-Lik: -479.388, Max-Change: 0.00020Iteration: 281, Log-Lik: -479.388, Max-Change: 0.00091Iteration: 282, Log-Lik: -479.388, Max-Change: 0.00148Iteration: 283, Log-Lik: -479.388, Max-Change: 0.00052Iteration: 284, Log-Lik: -479.388, Max-Change: 0.00042Iteration: 285, Log-Lik: -479.388, Max-Change: 0.00021Iteration: 286, Log-Lik: -479.388, Max-Change: 0.00019Iteration: 287, Log-Lik: -479.388, Max-Change: 0.00088Iteration: 288, Log-Lik: -479.388, Max-Change: 0.00295Iteration: 289, Log-Lik: -479.388, Max-Change: 0.00051Iteration: 290, Log-Lik: -479.388, Max-Change: 0.00051Iteration: 291, Log-Lik: -479.388, Max-Change: 0.00050Iteration: 292, Log-Lik: -479.388, Max-Change: 0.00133Iteration: 293, Log-Lik: -479.388, Max-Change: 0.00044Iteration: 294, Log-Lik: -479.388, Max-Change: 0.00021Iteration: 295, Log-Lik: -479.388, Max-Change: 0.00019Iteration: 296, Log-Lik: -479.388, Max-Change: 0.00088Iteration: 297, Log-Lik: -479.388, Max-Change: 0.00135Iteration: 298, Log-Lik: -479.388, Max-Change: 0.00046Iteration: 299, Log-Lik: -479.388, Max-Change: 0.00043Iteration: 300, Log-Lik: -479.388, Max-Change: 0.00022Iteration: 301, Log-Lik: -479.388, Max-Change: 0.00020Iteration: 302, Log-Lik: -479.388, Max-Change: 0.00089Iteration: 303, Log-Lik: -479.388, Max-Change: 0.00143Iteration: 304, Log-Lik: -479.387, Max-Change: 0.00051Iteration: 305, Log-Lik: -479.387, Max-Change: 0.00042Iteration: 306, Log-Lik: -479.387, Max-Change: 0.00021Iteration: 307, Log-Lik: -479.387, Max-Change: 0.00019Iteration: 308, Log-Lik: -479.387, Max-Change: 0.00086Iteration: 309, Log-Lik: -479.387, Max-Change: 0.00147Iteration: 310, Log-Lik: -479.387, Max-Change: 0.00051Iteration: 311, Log-Lik: -479.387, Max-Change: 0.00041Iteration: 312, Log-Lik: -479.387, Max-Change: 0.00021Iteration: 313, Log-Lik: -479.387, Max-Change: 0.00019Iteration: 314, Log-Lik: -479.387, Max-Change: 0.00085Iteration: 315, Log-Lik: -479.387, Max-Change: 0.00149Iteration: 316, Log-Lik: -479.387, Max-Change: 0.00050Iteration: 317, Log-Lik: -479.387, Max-Change: 0.00041Iteration: 318, Log-Lik: -479.387, Max-Change: 0.00021Iteration: 319, Log-Lik: -479.387, Max-Change: 0.00019Iteration: 320, Log-Lik: -479.387, Max-Change: 0.00085Iteration: 321, Log-Lik: -479.387, Max-Change: 0.00149Iteration: 322, Log-Lik: -479.387, Max-Change: 0.00050Iteration: 323, Log-Lik: -479.387, Max-Change: 0.00041Iteration: 324, Log-Lik: -479.387, Max-Change: 0.00021Iteration: 325, Log-Lik: -479.387, Max-Change: 0.00019Iteration: 326, Log-Lik: -479.387, Max-Change: 0.00085Iteration: 327, Log-Lik: -479.387, Max-Change: 0.00148Iteration: 328, Log-Lik: -479.387, Max-Change: 0.00050Iteration: 329, Log-Lik: -479.387, Max-Change: 0.00040Iteration: 330, Log-Lik: -479.387, Max-Change: 0.00021Iteration: 331, Log-Lik: -479.387, Max-Change: 0.00019Iteration: 332, Log-Lik: -479.387, Max-Change: 0.00084Iteration: 333, Log-Lik: -479.387, Max-Change: 0.00147Iteration: 334, Log-Lik: -479.387, Max-Change: 0.00049Iteration: 335, Log-Lik: -479.387, Max-Change: 0.00040Iteration: 336, Log-Lik: -479.387, Max-Change: 0.00020Iteration: 337, Log-Lik: -479.387, Max-Change: 0.00018Iteration: 338, Log-Lik: -479.387, Max-Change: 0.00084Iteration: 339, Log-Lik: -479.387, Max-Change: 0.00147Iteration: 340, Log-Lik: -479.387, Max-Change: 0.00049Iteration: 341, Log-Lik: -479.387, Max-Change: 0.00026Iteration: 342, Log-Lik: -479.387, Max-Change: 0.00022Iteration: 343, Log-Lik: -479.387, Max-Change: 0.00271Iteration: 344, Log-Lik: -479.387, Max-Change: 0.00048Iteration: 345, Log-Lik: -479.387, Max-Change: 0.00048Iteration: 346, Log-Lik: -479.386, Max-Change: 0.00142Iteration: 347, Log-Lik: -479.386, Max-Change: 0.00051Iteration: 348, Log-Lik: -479.386, Max-Change: 0.00023Iteration: 349, Log-Lik: -479.386, Max-Change: 0.00020Iteration: 350, Log-Lik: -479.386, Max-Change: 0.00018Iteration: 351, Log-Lik: -479.386, Max-Change: 0.00083Iteration: 352, Log-Lik: -479.386, Max-Change: 0.00134Iteration: 353, Log-Lik: -479.386, Max-Change: 0.00049Iteration: 354, Log-Lik: -479.386, Max-Change: 0.00043Iteration: 355, Log-Lik: -479.386, Max-Change: 0.00020Iteration: 356, Log-Lik: -479.386, Max-Change: 0.00019Iteration: 357, Log-Lik: -479.386, Max-Change: 0.00085Iteration: 358, Log-Lik: -479.386, Max-Change: 0.00151Iteration: 359, Log-Lik: -479.386, Max-Change: 0.00042Iteration: 360, Log-Lik: -479.386, Max-Change: 0.00023Iteration: 361, Log-Lik: -479.386, Max-Change: 0.00020Iteration: 362, Log-Lik: -479.386, Max-Change: 0.00089Iteration: 363, Log-Lik: -479.386, Max-Change: 0.00145Iteration: 364, Log-Lik: -479.386, Max-Change: 0.00049Iteration: 365, Log-Lik: -479.386, Max-Change: 0.00025Iteration: 366, Log-Lik: -479.386, Max-Change: 0.00021Iteration: 367, Log-Lik: -479.386, Max-Change: 0.00212Iteration: 368, Log-Lik: -479.386, Max-Change: 0.00051Iteration: 369, Log-Lik: -479.386, Max-Change: 0.00025Iteration: 370, Log-Lik: -479.386, Max-Change: 0.00020Iteration: 371, Log-Lik: -479.386, Max-Change: 0.00089Iteration: 372, Log-Lik: -479.386, Max-Change: 0.00154Iteration: 373, Log-Lik: -479.386, Max-Change: 0.00052Iteration: 374, Log-Lik: -479.386, Max-Change: 0.00025Iteration: 375, Log-Lik: -479.386, Max-Change: 0.00021Iteration: 376, Log-Lik: -479.386, Max-Change: 0.00326Iteration: 377, Log-Lik: -479.386, Max-Change: 0.00048Iteration: 378, Log-Lik: -479.386, Max-Change: 0.00046Iteration: 379, Log-Lik: -479.386, Max-Change: 0.00161Iteration: 380, Log-Lik: -479.386, Max-Change: 0.00022Iteration: 381, Log-Lik: -479.386, Max-Change: 0.00020Iteration: 382, Log-Lik: -479.386, Max-Change: 0.00156Iteration: 383, Log-Lik: -479.386, Max-Change: 0.00042Iteration: 384, Log-Lik: -479.386, Max-Change: 0.00023Iteration: 385, Log-Lik: -479.385, Max-Change: 0.00020Iteration: 386, Log-Lik: -479.385, Max-Change: 0.00089Iteration: 387, Log-Lik: -479.385, Max-Change: 0.00131Iteration: 388, Log-Lik: -479.385, Max-Change: 0.00048Iteration: 389, Log-Lik: -479.385, Max-Change: 0.00024Iteration: 390, Log-Lik: -479.385, Max-Change: 0.00021Iteration: 391, Log-Lik: -479.385, Max-Change: 0.00250Iteration: 392, Log-Lik: -479.385, Max-Change: 0.00048Iteration: 393, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 394, Log-Lik: -479.385, Max-Change: 0.00022Iteration: 395, Log-Lik: -479.385, Max-Change: 0.00098Iteration: 396, Log-Lik: -479.385, Max-Change: 0.00133Iteration: 397, Log-Lik: -479.385, Max-Change: 0.00050Iteration: 398, Log-Lik: -479.385, Max-Change: 0.00026Iteration: 399, Log-Lik: -479.385, Max-Change: 0.00022Iteration: 400, Log-Lik: -479.385, Max-Change: 0.00228Iteration: 401, Log-Lik: -479.385, Max-Change: 0.00056Iteration: 402, Log-Lik: -479.385, Max-Change: 0.00045Iteration: 403, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 404, Log-Lik: -479.385, Max-Change: 0.00021Iteration: 405, Log-Lik: -479.385, Max-Change: 0.00095Iteration: 406, Log-Lik: -479.385, Max-Change: 0.00139Iteration: 407, Log-Lik: -479.385, Max-Change: 0.00026Iteration: 408, Log-Lik: -479.385, Max-Change: 0.00022Iteration: 409, Log-Lik: -479.385, Max-Change: 0.00127Iteration: 410, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 411, Log-Lik: -479.385, Max-Change: 0.00021Iteration: 412, Log-Lik: -479.385, Max-Change: 0.00129Iteration: 413, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 414, Log-Lik: -479.385, Max-Change: 0.00022Iteration: 415, Log-Lik: -479.385, Max-Change: 0.00100Iteration: 416, Log-Lik: -479.385, Max-Change: 0.00022Iteration: 417, Log-Lik: -479.385, Max-Change: 0.00019Iteration: 418, Log-Lik: -479.385, Max-Change: 0.00141Iteration: 419, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 420, Log-Lik: -479.385, Max-Change: 0.00021Iteration: 421, Log-Lik: -479.385, Max-Change: 0.00140Iteration: 422, Log-Lik: -479.385, Max-Change: 0.00025Iteration: 423, Log-Lik: -479.385, Max-Change: 0.00021Iteration: 424, Log-Lik: -479.385, Max-Change: 0.00118Iteration: 425, Log-Lik: -479.385, Max-Change: 0.00023Iteration: 426, Log-Lik: -479.385, Max-Change: 0.00020Iteration: 427, Log-Lik: -479.385, Max-Change: 0.00139Iteration: 428, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 429, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 430, Log-Lik: -479.384, Max-Change: 0.00124Iteration: 431, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 432, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 433, Log-Lik: -479.384, Max-Change: 0.00131Iteration: 434, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 435, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 436, Log-Lik: -479.384, Max-Change: 0.00134Iteration: 437, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 438, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 439, Log-Lik: -479.384, Max-Change: 0.00155Iteration: 440, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 441, Log-Lik: -479.384, Max-Change: 0.00018Iteration: 442, Log-Lik: -479.384, Max-Change: 0.00250Iteration: 443, Log-Lik: -479.384, Max-Change: 0.00049Iteration: 444, Log-Lik: -479.384, Max-Change: 0.00024Iteration: 445, Log-Lik: -479.384, Max-Change: 0.00022Iteration: 446, Log-Lik: -479.384, Max-Change: 0.00096Iteration: 447, Log-Lik: -479.384, Max-Change: 0.00132Iteration: 448, Log-Lik: -479.384, Max-Change: 0.00052Iteration: 449, Log-Lik: -479.384, Max-Change: 0.00024Iteration: 450, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 451, Log-Lik: -479.384, Max-Change: 0.00166Iteration: 452, Log-Lik: -479.384, Max-Change: 0.00048Iteration: 453, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 454, Log-Lik: -479.384, Max-Change: 0.00018Iteration: 455, Log-Lik: -479.384, Max-Change: 0.00082Iteration: 456, Log-Lik: -479.384, Max-Change: 0.00143Iteration: 457, Log-Lik: -479.384, Max-Change: 0.00045Iteration: 458, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 459, Log-Lik: -479.384, Max-Change: 0.00021Iteration: 460, Log-Lik: -479.384, Max-Change: 0.00317Iteration: 461, Log-Lik: -479.384, Max-Change: 0.00052Iteration: 462, Log-Lik: -479.384, Max-Change: 0.00048Iteration: 463, Log-Lik: -479.384, Max-Change: 0.00012Iteration: 464, Log-Lik: -479.384, Max-Change: 0.00135Iteration: 465, Log-Lik: -479.384, Max-Change: 0.00026Iteration: 466, Log-Lik: -479.384, Max-Change: 0.00025Iteration: 467, Log-Lik: -479.384, Max-Change: 0.00022Iteration: 468, Log-Lik: -479.384, Max-Change: 0.00095Iteration: 469, Log-Lik: -479.384, Max-Change: 0.00099Iteration: 470, Log-Lik: -479.384, Max-Change: 0.00019Iteration: 471, Log-Lik: -479.383, Max-Change: 0.00083Iteration: 472, Log-Lik: -479.383, Max-Change: 0.00012Iteration: 473, Log-Lik: -479.383, Max-Change: 0.00144Iteration: 474, Log-Lik: -479.383, Max-Change: 0.00025Iteration: 475, Log-Lik: -479.383, Max-Change: 0.00025Iteration: 476, Log-Lik: -479.383, Max-Change: 0.00021Iteration: 477, Log-Lik: -479.383, Max-Change: 0.00092Iteration: 478, Log-Lik: -479.383, Max-Change: 0.00329Iteration: 479, Log-Lik: -479.383, Max-Change: 0.00070Iteration: 480, Log-Lik: -479.383, Max-Change: 0.00048Iteration: 481, Log-Lik: -479.383, Max-Change: 0.00015Iteration: 482, Log-Lik: -479.383, Max-Change: 0.00071Iteration: 483, Log-Lik: -479.383, Max-Change: 0.00057Iteration: 484, Log-Lik: -479.383, Max-Change: 0.00026Iteration: 485, Log-Lik: -479.383, Max-Change: 0.00011Iteration: 486, Log-Lik: -479.383, Max-Change: 0.00267Iteration: 487, Log-Lik: -479.383, Max-Change: 0.00046Iteration: 488, Log-Lik: -479.383, Max-Change: 0.00021Iteration: 489, Log-Lik: -479.383, Max-Change: 0.00019Iteration: 490, Log-Lik: -479.383, Max-Change: 0.00084Iteration: 491, Log-Lik: -479.383, Max-Change: 0.00020Iteration: 492, Log-Lik: -479.383, Max-Change: 0.00090Iteration: 493, Log-Lik: -479.383, Max-Change: 0.00013Iteration: 494, Log-Lik: -479.383, Max-Change: 0.00138Iteration: 495, Log-Lik: -479.383, Max-Change: 0.00025Iteration: 496, Log-Lik: -479.383, Max-Change: 0.00025Iteration: 497, Log-Lik: -479.383, Max-Change: 0.00021Iteration: 498, Log-Lik: -479.383, Max-Change: 0.00092Iteration: 499, Log-Lik: -479.383, Max-Change: 0.00152Iteration: 500, Log-Lik: -479.383, Max-Change: 0.00025

#> EM cycles terminated after 500 iterations.

out <- model_goodness_for_algo_poly(mod$model, num=1)
out

#> $xy
#>         x    y
#> [1,] 0.00 0.26
#> [2,] 0.25 0.44
#> [3,] 0.50 0.66
#> [4,] 0.75 0.85
#> [5,] 1.00 1.00
#> 
#> $auc
#> [1] 0.645
#> 
#> $mse
#> [1] 5.17
#>