Computes the actual and predicted effectiveness of the collection of algorithms.

This function computes the actual and predicted effectiveness of the collection of algorithms for different tolerance values.

effectiveness_crm(mod)

Arguments

mod
A fitted mirt model using the function irtmodel or R package mirt.

Value

A list with the following components:

effectivenessAUC

The area under the actual and predicted effectiveness curves.

actcurves

The x,y coodinates for the actual effectiveness curves for each algorithm.

#'
prdcurves

The x,y coodinates for the predicted effectiveness curves for each algorithm.

Examples

set.seed(1)# \donttest{
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)

#> Warning: NaNs produced

#> Warning: NaNs produced

out <- effectiveness_crm(mod$model)
out

#> $effectivenessAUC
#>       Actual Predicted
#> x1 0.5128283 0.5174747
#> x2 0.5129293 0.7654545
#> x3 0.4292929 0.2778283
#> 
#> $actcurves
#>                 x   x1   x2   x3
#>   [1,] 0.00000000 0.00 0.00 0.00
#>   [2,] 0.01010101 0.01 0.02 0.00
#>   [3,] 0.02020202 0.01 0.03 0.01
#>   [4,] 0.03030303 0.01 0.04 0.01
#>   [5,] 0.04040404 0.02 0.04 0.02
#>   [6,] 0.05050505 0.02 0.04 0.04
#>   [7,] 0.06060606 0.03 0.05 0.06
#>   [8,] 0.07070707 0.04 0.05 0.08
#>   [9,] 0.08080808 0.04 0.08 0.10
#>  [10,] 0.09090909 0.05 0.08 0.11
#>  [11,] 0.10101010 0.06 0.09 0.11
#>  [12,] 0.11111111 0.08 0.11 0.11
#>  [13,] 0.12121212 0.08 0.13 0.12
#>  [14,] 0.13131313 0.10 0.14 0.12
#>  [15,] 0.14141414 0.13 0.15 0.12
#>  [16,] 0.15151515 0.13 0.16 0.13
#>  [17,] 0.16161616 0.13 0.18 0.13
#>  [18,] 0.17171717 0.14 0.18 0.13
#>  [19,] 0.18181818 0.16 0.18 0.13
#>  [20,] 0.19191919 0.17 0.18 0.14
#>  [21,] 0.20202020 0.17 0.18 0.14
#>  [22,] 0.21212121 0.19 0.18 0.14
#>  [23,] 0.22222222 0.21 0.19 0.15
#>  [24,] 0.23232323 0.24 0.20 0.16
#>  [25,] 0.24242424 0.26 0.20 0.18
#>  [26,] 0.25252525 0.27 0.22 0.19
#>  [27,] 0.26262626 0.27 0.24 0.19
#>  [28,] 0.27272727 0.28 0.25 0.19
#>  [29,] 0.28282828 0.29 0.28 0.19
#>  [30,] 0.29292929 0.32 0.30 0.20
#>  [31,] 0.30303030 0.32 0.30 0.21
#>  [32,] 0.31313131 0.33 0.30 0.21
#>  [33,] 0.32323232 0.34 0.31 0.24
#>  [34,] 0.33333333 0.34 0.31 0.24
#>  [35,] 0.34343434 0.37 0.31 0.24
#>  [36,] 0.35353535 0.39 0.32 0.24
#>  [37,] 0.36363636 0.41 0.35 0.25
#>  [38,] 0.37373737 0.41 0.37 0.26
#>  [39,] 0.38383838 0.42 0.37 0.27
#>  [40,] 0.39393939 0.42 0.38 0.27
#>  [41,] 0.40404040 0.43 0.40 0.28
#>  [42,] 0.41414141 0.44 0.44 0.28
#>  [43,] 0.42424242 0.44 0.44 0.28
#>  [44,] 0.43434343 0.45 0.45 0.29
#>  [45,] 0.44444444 0.45 0.46 0.30
#>  [46,] 0.45454545 0.46 0.47 0.31
#>  [47,] 0.46464646 0.46 0.48 0.31
#>  [48,] 0.47474747 0.46 0.48 0.31
#>  [49,] 0.48484848 0.47 0.50 0.33
#>  [50,] 0.49494949 0.48 0.51 0.37
#>  [51,] 0.50505051 0.48 0.54 0.38
#>  [52,] 0.51515152 0.50 0.55 0.39
#>  [53,] 0.52525253 0.51 0.56 0.40
#>  [54,] 0.53535354 0.55 0.57 0.42
#>  [55,] 0.54545455 0.55 0.57 0.43
#>  [56,] 0.55555556 0.57 0.60 0.44
#>  [57,] 0.56565657 0.57 0.62 0.44
#>  [58,] 0.57575758 0.59 0.63 0.44
#>  [59,] 0.58585859 0.59 0.63 0.46
#>  [60,] 0.59595960 0.61 0.64 0.48
#>  [61,] 0.60606061 0.62 0.64 0.48
#>  [62,] 0.61616162 0.63 0.65 0.49
#>  [63,] 0.62626263 0.68 0.67 0.50
#>  [64,] 0.63636364 0.69 0.67 0.53
#>  [65,] 0.64646465 0.69 0.67 0.53
#>  [66,] 0.65656566 0.69 0.69 0.53
#>  [67,] 0.66666667 0.71 0.69 0.54
#>  [68,] 0.67676768 0.74 0.71 0.55
#>  [69,] 0.68686869 0.75 0.72 0.58
#>  [70,] 0.69696970 0.76 0.73 0.59
#>  [71,] 0.70707071 0.76 0.73 0.61
#>  [72,] 0.71717172 0.77 0.74 0.62
#>  [73,] 0.72727273 0.77 0.76 0.62
#>  [74,] 0.73737374 0.77 0.78 0.63
#>  [75,] 0.74747475 0.79 0.78 0.66
#>  [76,] 0.75757576 0.80 0.79 0.69
#>  [77,] 0.76767677 0.82 0.79 0.70
#>  [78,] 0.77777778 0.83 0.80 0.70
#>  [79,] 0.78787879 0.83 0.81 0.72
#>  [80,] 0.79797980 0.84 0.83 0.74
#>  [81,] 0.80808081 0.87 0.84 0.74
#>  [82,] 0.81818182 0.87 0.85 0.75
#>  [83,] 0.82828283 0.88 0.87 0.77
#>  [84,] 0.83838384 0.89 0.89 0.77
#>  [85,] 0.84848485 0.89 0.89 0.81
#>  [86,] 0.85858586 0.89 0.89 0.82
#>  [87,] 0.86868687 0.90 0.90 0.83
#>  [88,] 0.87878788 0.90 0.90 0.88
#>  [89,] 0.88888889 0.92 0.91 0.89
#>  [90,] 0.89898990 0.92 0.92 0.90
#>  [91,] 0.90909091 0.93 0.95 0.93
#>  [92,] 0.91919192 0.94 0.95 0.94
#>  [93,] 0.92929293 0.95 0.95 0.94
#>  [94,] 0.93939394 0.96 0.97 0.94
#>  [95,] 0.94949495 0.97 0.98 0.96
#>  [96,] 0.95959596 0.98 0.98 0.97
#>  [97,] 0.96969697 0.98 0.98 0.99
#>  [98,] 0.97979798 0.98 0.99 0.99
#>  [99,] 0.98989899 0.99 0.99 1.00
#> [100,] 1.00000000 1.00 1.00 1.00
#> 
#> $prdcurves
#>                 x   x1   x2   x3
#>   [1,] 0.00000000 0.00 0.00 0.00
#>   [2,] 0.01010101 0.00 0.48 0.01
#>   [3,] 0.02020202 0.00 0.50 0.01
#>   [4,] 0.03030303 0.01 0.52 0.02
#>   [5,] 0.04040404 0.01 0.56 0.02
#>   [6,] 0.05050505 0.01 0.56 0.02
#>   [7,] 0.06060606 0.01 0.57 0.02
#>   [8,] 0.07070707 0.02 0.61 0.03
#>   [9,] 0.08080808 0.02 0.63 0.03
#>  [10,] 0.09090909 0.02 0.64 0.03
#>  [11,] 0.10101010 0.02 0.65 0.04
#>  [12,] 0.11111111 0.02 0.66 0.04
#>  [13,] 0.12121212 0.03 0.66 0.06
#>  [14,] 0.13131313 0.03 0.66 0.06
#>  [15,] 0.14141414 0.04 0.66 0.06
#>  [16,] 0.15151515 0.06 0.66 0.06
#>  [17,] 0.16161616 0.06 0.68 0.06
#>  [18,] 0.17171717 0.07 0.69 0.07
#>  [19,] 0.18181818 0.07 0.69 0.07
#>  [20,] 0.19191919 0.07 0.69 0.07
#>  [21,] 0.20202020 0.10 0.69 0.07
#>  [22,] 0.21212121 0.10 0.69 0.07
#>  [23,] 0.22222222 0.12 0.69 0.09
#>  [24,] 0.23232323 0.13 0.71 0.09
#>  [25,] 0.24242424 0.16 0.71 0.10
#>  [26,] 0.25252525 0.17 0.71 0.10
#>  [27,] 0.26262626 0.17 0.71 0.10
#>  [28,] 0.27272727 0.22 0.71 0.10
#>  [29,] 0.28282828 0.23 0.72 0.10
#>  [30,] 0.29292929 0.26 0.72 0.12
#>  [31,] 0.30303030 0.28 0.72 0.12
#>  [32,] 0.31313131 0.30 0.72 0.13
#>  [33,] 0.32323232 0.31 0.72 0.13
#>  [34,] 0.33333333 0.33 0.73 0.14
#>  [35,] 0.34343434 0.33 0.74 0.14
#>  [36,] 0.35353535 0.34 0.74 0.17
#>  [37,] 0.36363636 0.36 0.74 0.17
#>  [38,] 0.37373737 0.37 0.75 0.17
#>  [39,] 0.38383838 0.37 0.75 0.17
#>  [40,] 0.39393939 0.39 0.75 0.17
#>  [41,] 0.40404040 0.40 0.75 0.17
#>  [42,] 0.41414141 0.41 0.75 0.18
#>  [43,] 0.42424242 0.43 0.76 0.21
#>  [44,] 0.43434343 0.44 0.76 0.22
#>  [45,] 0.44444444 0.46 0.76 0.23
#>  [46,] 0.45454545 0.48 0.76 0.23
#>  [47,] 0.46464646 0.50 0.77 0.23
#>  [48,] 0.47474747 0.50 0.77 0.26
#>  [49,] 0.48484848 0.50 0.77 0.26
#>  [50,] 0.49494949 0.51 0.78 0.26
#>  [51,] 0.50505051 0.52 0.78 0.27
#>  [52,] 0.51515152 0.54 0.78 0.28
#>  [53,] 0.52525253 0.56 0.79 0.29
#>  [54,] 0.53535354 0.56 0.79 0.29
#>  [55,] 0.54545455 0.57 0.79 0.30
#>  [56,] 0.55555556 0.61 0.79 0.30
#>  [57,] 0.56565657 0.64 0.80 0.31
#>  [58,] 0.57575758 0.65 0.80 0.31
#>  [59,] 0.58585859 0.66 0.81 0.32
#>  [60,] 0.59595960 0.66 0.82 0.33
#>  [61,] 0.60606061 0.69 0.82 0.33
#>  [62,] 0.61616162 0.69 0.82 0.33
#>  [63,] 0.62626263 0.70 0.83 0.33
#>  [64,] 0.63636364 0.72 0.83 0.34
#>  [65,] 0.64646465 0.72 0.83 0.34
#>  [66,] 0.65656566 0.74 0.83 0.35
#>  [67,] 0.66666667 0.76 0.83 0.36
#>  [68,] 0.67676768 0.77 0.83 0.37
#>  [69,] 0.68686869 0.79 0.84 0.37
#>  [70,] 0.69696970 0.82 0.84 0.37
#>  [71,] 0.70707071 0.83 0.84 0.37
#>  [72,] 0.71717172 0.84 0.84 0.37
#>  [73,] 0.72727273 0.85 0.84 0.38
#>  [74,] 0.73737374 0.85 0.85 0.40
#>  [75,] 0.74747475 0.85 0.85 0.40
#>  [76,] 0.75757576 0.87 0.85 0.40
#>  [77,] 0.76767677 0.87 0.85 0.42
#>  [78,] 0.77777778 0.87 0.85 0.43
#>  [79,] 0.78787879 0.89 0.85 0.43
#>  [80,] 0.79797980 0.92 0.85 0.44
#>  [81,] 0.80808081 0.93 0.85 0.46
#>  [82,] 0.81818182 0.94 0.85 0.47
#>  [83,] 0.82828283 0.95 0.85 0.48
#>  [84,] 0.83838384 0.95 0.86 0.50
#>  [85,] 0.84848485 0.96 0.86 0.50
#>  [86,] 0.85858586 0.96 0.87 0.50
#>  [87,] 0.86868687 0.96 0.87 0.50
#>  [88,] 0.87878788 0.97 0.87 0.51
#>  [89,] 0.88888889 0.97 0.87 0.52
#>  [90,] 0.89898990 0.97 0.87 0.56
#>  [91,] 0.90909091 0.98 0.87 0.56
#>  [92,] 0.91919192 0.98 0.87 0.57
#>  [93,] 0.92929293 0.98 0.87 0.62
#>  [94,] 0.93939394 0.99 0.89 0.65
#>  [95,] 0.94949495 0.99 0.90 0.66
#>  [96,] 0.95959596 1.00 0.93 0.69
#>  [97,] 0.96969697 1.00 0.93 0.72
#>  [98,] 0.97979798 1.00 0.93 0.76
#>  [99,] 0.98989899 1.00 0.95 0.83
#> [100,] 1.00000000 1.00 0.96 0.87
#> 

# }