Computes the goodness of IRT model for all algorithms.

This function computes the goodness of the IRT model for all algorithms for different goodness tolerances.

model_goodness_crm(mod)

Arguments

mod

A fitted mirt model using the function irtmodel or R package mirt.

Value

A list with the following components:

goodnessAUC

The area under the model goodness curve for each algorithm.

curves

The x,y coodinates for the model goodness curves for each algorithm.

Examples

# \donttest{
set.seed(1)
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)
#> Warning: NaNs produced
#> Warning: NaNs produced
out <- model_goodness_crm(mod$model)
out
#> $goodnessAUC
#>         [,1]
#> x1 0.9253535
#> x2 0.5859596
#> x3 0.7127273
#> 
#> $mse
#>             [,1]
#> [1,] 0.008553221
#> [2,] 0.238807237
#> [3,] 0.125930464
#> 
#> $curves
#>                 x   x1   x2   x3
#>   [1,] 0.00000000 0.00 0.00 0.00
#>   [2,] 0.01010101 0.11 0.04 0.01
#>   [3,] 0.02020202 0.18 0.05 0.03
#>   [4,] 0.03030303 0.21 0.05 0.05
#>   [5,] 0.04040404 0.32 0.05 0.07
#>   [6,] 0.05050505 0.38 0.06 0.13
#>   [7,] 0.06060606 0.45 0.06 0.16
#>   [8,] 0.07070707 0.55 0.06 0.18
#>   [9,] 0.08080808 0.59 0.09 0.18
#>  [10,] 0.09090909 0.66 0.09 0.19
#>  [11,] 0.10101010 0.75 0.11 0.20
#>  [12,] 0.11111111 0.77 0.13 0.22
#>  [13,] 0.12121212 0.83 0.19 0.22
#>  [14,] 0.13131313 0.86 0.21 0.25
#>  [15,] 0.14141414 0.89 0.24 0.28
#>  [16,] 0.15151515 0.92 0.24 0.33
#>  [17,] 0.16161616 0.94 0.26 0.34
#>  [18,] 0.17171717 0.94 0.27 0.36
#>  [19,] 0.18181818 0.94 0.30 0.38
#>  [20,] 0.19191919 0.96 0.31 0.38
#>  [21,] 0.20202020 0.98 0.31 0.41
#>  [22,] 0.21212121 0.98 0.32 0.43
#>  [23,] 0.22222222 0.98 0.35 0.46
#>  [24,] 0.23232323 0.99 0.35 0.47
#>  [25,] 0.24242424 0.99 0.35 0.53
#>  [26,] 0.25252525 0.99 0.36 0.55
#>  [27,] 0.26262626 0.99 0.37 0.58
#>  [28,] 0.27272727 0.99 0.38 0.60
#>  [29,] 0.28282828 0.99 0.38 0.62
#>  [30,] 0.29292929 0.99 0.39 0.62
#>  [31,] 0.30303030 0.99 0.39 0.62
#>  [32,] 0.31313131 1.00 0.39 0.64
#>  [33,] 0.32323232 1.00 0.40 0.67
#>  [34,] 0.33333333 1.00 0.41 0.69
#>  [35,] 0.34343434 1.00 0.42 0.69
#>  [36,] 0.35353535 1.00 0.42 0.70
#>  [37,] 0.36363636 1.00 0.44 0.70
#>  [38,] 0.37373737 1.00 0.44 0.71
#>  [39,] 0.38383838 1.00 0.47 0.72
#>  [40,] 0.39393939 1.00 0.48 0.73
#>  [41,] 0.40404040 1.00 0.51 0.75
#>  [42,] 0.41414141 1.00 0.53 0.76
#>  [43,] 0.42424242 1.00 0.53 0.76
#>  [44,] 0.43434343 1.00 0.54 0.79
#>  [45,] 0.44444444 1.00 0.55 0.80
#>  [46,] 0.45454545 1.00 0.56 0.80
#>  [47,] 0.46464646 1.00 0.58 0.81
#>  [48,] 0.47474747 1.00 0.59 0.81
#>  [49,] 0.48484848 1.00 0.59 0.81
#>  [50,] 0.49494949 1.00 0.61 0.82
#>  [51,] 0.50505051 1.00 0.62 0.82
#>  [52,] 0.51515152 1.00 0.63 0.83
#>  [53,] 0.52525253 1.00 0.66 0.83
#>  [54,] 0.53535354 1.00 0.67 0.83
#>  [55,] 0.54545455 1.00 0.67 0.83
#>  [56,] 0.55555556 1.00 0.69 0.84
#>  [57,] 0.56565657 1.00 0.70 0.85
#>  [58,] 0.57575758 1.00 0.72 0.87
#>  [59,] 0.58585859 1.00 0.73 0.87
#>  [60,] 0.59595960 1.00 0.73 0.87
#>  [61,] 0.60606061 1.00 0.73 0.87
#>  [62,] 0.61616162 1.00 0.73 0.88
#>  [63,] 0.62626263 1.00 0.74 0.90
#>  [64,] 0.63636364 1.00 0.74 0.90
#>  [65,] 0.64646465 1.00 0.74 0.90
#>  [66,] 0.65656566 1.00 0.75 0.91
#>  [67,] 0.66666667 1.00 0.76 0.91
#>  [68,] 0.67676768 1.00 0.78 0.92
#>  [69,] 0.68686869 1.00 0.81 0.93
#>  [70,] 0.69696970 1.00 0.81 0.93
#>  [71,] 0.70707071 1.00 0.82 0.93
#>  [72,] 0.71717172 1.00 0.82 0.93
#>  [73,] 0.72727273 1.00 0.85 0.94
#>  [74,] 0.73737374 1.00 0.86 0.94
#>  [75,] 0.74747475 1.00 0.86 0.97
#>  [76,] 0.75757576 1.00 0.86 0.98
#>  [77,] 0.76767677 1.00 0.86 0.98
#>  [78,] 0.77777778 1.00 0.87 0.98
#>  [79,] 0.78787879 1.00 0.89 0.98
#>  [80,] 0.79797980 1.00 0.90 0.98
#>  [81,] 0.80808081 1.00 0.90 0.98
#>  [82,] 0.81818182 1.00 0.91 0.98
#>  [83,] 0.82828283 1.00 0.91 0.99
#>  [84,] 0.83838384 1.00 0.92 1.00
#>  [85,] 0.84848485 1.00 0.92 1.00
#>  [86,] 0.85858586 1.00 0.92 1.00
#>  [87,] 0.86868687 1.00 0.93 1.00
#>  [88,] 0.87878788 1.00 0.95 1.00
#>  [89,] 0.88888889 1.00 0.96 1.00
#>  [90,] 0.89898990 1.00 0.97 1.00
#>  [91,] 0.90909091 1.00 1.00 1.00
#>  [92,] 0.91919192 1.00 1.00 1.00
#>  [93,] 0.92929293 1.00 1.00 1.00
#>  [94,] 0.93939394 1.00 1.00 1.00
#>  [95,] 0.94949495 1.00 1.00 1.00
#>  [96,] 0.95959596 1.00 1.00 1.00
#>  [97,] 0.96969697 1.00 1.00 1.00
#>  [98,] 0.97979798 1.00 1.00 1.00
#>  [99,] 0.98989899 1.00 1.00 1.00
#> [100,] 1.00000000 1.00 1.00 1.00
#> 
# }