Computes the goodness of IRT model for a given algorithm.

This function computes the goodness of the IRT model for a given algorithm for different goodness tolerances.

model_goodness_for_algo_crm(mod, num = 1)

Arguments

mod

A fitted mirt model using the function irtmodel or R package mirt.

num

The algorithm number, for which the goodness of the IRT model is computed.

Value

A list with the following components:

xy

The x values denote the goodness tolerances. The y values denote the model goodness.

auc

The area under the model goodness curve.

residuals

The different between actual and fitted performance values.

Examples

# \donttest{
set.seed(1)
x1 <- runif(100)
x2 <- runif(100)
x3 <- runif(100)
X <- cbind.data.frame(x1, x2, x3)
max_item <- rep(1,3)
min_item <- rep(0,3)
mod <- cirtmodel(X, max.item=max_item, min.item=min_item)
#> Warning: NaNs produced
#> Warning: NaNs produced
out <- model_goodness_for_algo_crm(mod$model, num=1)
out
#> $xy
#>                 x    y
#>   [1,] 0.00000000 0.00
#>   [2,] 0.01010101 0.08
#>   [3,] 0.02020202 0.16
#>   [4,] 0.03030303 0.21
#>   [5,] 0.04040404 0.30
#>   [6,] 0.05050505 0.39
#>   [7,] 0.06060606 0.40
#>   [8,] 0.07070707 0.52
#>   [9,] 0.08080808 0.56
#>  [10,] 0.09090909 0.61
#>  [11,] 0.10101010 0.69
#>  [12,] 0.11111111 0.75
#>  [13,] 0.12121212 0.81
#>  [14,] 0.13131313 0.84
#>  [15,] 0.14141414 0.85
#>  [16,] 0.15151515 0.89
#>  [17,] 0.16161616 0.91
#>  [18,] 0.17171717 0.93
#>  [19,] 0.18181818 0.93
#>  [20,] 0.19191919 0.95
#>  [21,] 0.20202020 0.96
#>  [22,] 0.21212121 0.97
#>  [23,] 0.22222222 0.97
#>  [24,] 0.23232323 0.98
#>  [25,] 0.24242424 0.99
#>  [26,] 0.25252525 0.99
#>  [27,] 0.26262626 0.99
#>  [28,] 0.27272727 0.99
#>  [29,] 0.28282828 0.99
#>  [30,] 0.29292929 0.99
#>  [31,] 0.30303030 0.99
#>  [32,] 0.31313131 0.99
#>  [33,] 0.32323232 0.99
#>  [34,] 0.33333333 0.99
#>  [35,] 0.34343434 1.00
#>  [36,] 0.35353535 1.00
#>  [37,] 0.36363636 1.00
#>  [38,] 0.37373737 1.00
#>  [39,] 0.38383838 1.00
#>  [40,] 0.39393939 1.00
#>  [41,] 0.40404040 1.00
#>  [42,] 0.41414141 1.00
#>  [43,] 0.42424242 1.00
#>  [44,] 0.43434343 1.00
#>  [45,] 0.44444444 1.00
#>  [46,] 0.45454545 1.00
#>  [47,] 0.46464646 1.00
#>  [48,] 0.47474747 1.00
#>  [49,] 0.48484848 1.00
#>  [50,] 0.49494949 1.00
#>  [51,] 0.50505051 1.00
#>  [52,] 0.51515152 1.00
#>  [53,] 0.52525253 1.00
#>  [54,] 0.53535354 1.00
#>  [55,] 0.54545455 1.00
#>  [56,] 0.55555556 1.00
#>  [57,] 0.56565657 1.00
#>  [58,] 0.57575758 1.00
#>  [59,] 0.58585859 1.00
#>  [60,] 0.59595960 1.00
#>  [61,] 0.60606061 1.00
#>  [62,] 0.61616162 1.00
#>  [63,] 0.62626263 1.00
#>  [64,] 0.63636364 1.00
#>  [65,] 0.64646465 1.00
#>  [66,] 0.65656566 1.00
#>  [67,] 0.66666667 1.00
#>  [68,] 0.67676768 1.00
#>  [69,] 0.68686869 1.00
#>  [70,] 0.69696970 1.00
#>  [71,] 0.70707071 1.00
#>  [72,] 0.71717172 1.00
#>  [73,] 0.72727273 1.00
#>  [74,] 0.73737374 1.00
#>  [75,] 0.74747475 1.00
#>  [76,] 0.75757576 1.00
#>  [77,] 0.76767677 1.00
#>  [78,] 0.77777778 1.00
#>  [79,] 0.78787879 1.00
#>  [80,] 0.79797980 1.00
#>  [81,] 0.80808081 1.00
#>  [82,] 0.81818182 1.00
#>  [83,] 0.82828283 1.00
#>  [84,] 0.83838384 1.00
#>  [85,] 0.84848485 1.00
#>  [86,] 0.85858586 1.00
#>  [87,] 0.86868687 1.00
#>  [88,] 0.87878788 1.00
#>  [89,] 0.88888889 1.00
#>  [90,] 0.89898990 1.00
#>  [91,] 0.90909091 1.00
#>  [92,] 0.91919192 1.00
#>  [93,] 0.92929293 1.00
#>  [94,] 0.93939394 1.00
#>  [95,] 0.94949495 1.00
#>  [96,] 0.95959596 1.00
#>  [97,] 0.96969697 1.00
#>  [98,] 0.97979798 1.00
#>  [99,] 0.98989899 1.00
#> [100,] 1.00000000 1.00
#> 
#> $auc
#> [1] 0.919798
#> 
#> $mse
#> [1] 0.00981765
#> 
#> $residuals
#>   [1] 0.0666564104 0.0005122839 0.0198855716 0.0653663279 0.0605079228
#>   [6] 0.0979209431 0.0846922540 0.1248511135 0.0278886544 0.1139741102
#>  [11] 0.3354224011 0.0452819937 0.0081519399 0.1985013063 0.0407511023
#>  [16] 0.0929549593 0.0171364067 0.0319324951 0.2089608444 0.0877170930
#>  [21] 0.0455578045 0.0460699389 0.0663862995 0.0916320207 0.2316645863
#>  [26] 0.0610482614 0.0368224545 0.1091784043 0.1115118601 0.1636522211
#>  [31] 0.0055264591 0.1490217144 0.0642007216 0.1407329603 0.0620067679
#>  [36] 0.0005017591 0.0446748238 0.1153611385 0.0039561304 0.0459377211
#>  [41] 0.0983730126 0.0118277194 0.0078187054 0.1002110744 0.0132717574
#>  [46] 0.1282707092 0.0278672117 0.0380982843 0.1015702353 0.0873371712
#>  [51] 0.0900281395 0.0102289839 0.0262824215 0.1162351167 0.0915372537
#>  [56] 0.0338461152 0.0388916385 0.0324527647 0.0755177529 0.1577436915
#>  [61] 0.0906378665 0.0948819533 0.0100235504 0.1603679295 0.0339384333
#>  [66] 0.0283326642 0.1503108872 0.0380641418 0.1245991718 0.1473854069
#>  [71] 0.0406667902 0.1377424806 0.0856733760 0.0028819103 0.0166537202
#>  [76] 0.1027734177 0.1120204463 0.0117253457 0.0685125907 0.0359452580
#>  [81] 0.1485388799 0.0727761558 0.1852726895 0.0596275080 0.1823811750
#>  [86] 0.0215683225 0.0647211993 0.0689070119 0.0698752739 0.0730441224
#>  [91] 0.0682521664 0.0434055800 0.1011323141 0.0771167530 0.0847742984
#>  [96] 0.0451310460 0.1022423493 0.0092621857 0.2296637601 0.1123234708
#> 
# }